separate appointment details to its own component
